+++ Tests/SentryTests/Integrations/SessionReplay/SentryReplayTypeTests.swift
@@ -1,4 +1,4 @@
-@testable import Sentry
+@_spi(Private) @testable import Sentry
 import XCTest
 
 class SentryReplayTypeTests: XCTestCase {